Fixed usnigned int comparison.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@29650 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ display/src/EVEVisualization.cc
@@ -352,7 +352,7 @@ void EVEVisualization::addTrack(const Belle2::Track* belle2Track)
       if (prevFittedState != NULL) {
 
         TEvePathMark::EType_e markType = TEvePathMark::kReference;
-        if (hitCounter + 1 == numpoints) //track should stop here.
+        if (hitCounter + 1 == static_cast<int>(numpoints)) //track should stop here.
           markType = TEvePathMark::kDecay;
 
         // Kalman: non-null prevFi ensures that the previous fitter info was also KalmanFitterInfo
